ACALF/requirements.txt
detectron2==0.6
fvcore==0.1.5
matplotlib==3.8.0
numpy==1.24.1
opencv_python==4.8.0.76
Pillow==10.4.0
tensorboardX==2.6.2
timm==0.9.5
torch==2.0.1+cu118
torchvision==0.15.2+cu118
